refacto Profil Page
